+++ app/src/androidTest/java/app/paseico/MarketplaceFragmentTest.java
@@ -1,51 +1,23 @@
 package app.paseico;
 
-import android.util.Log;
-import android.widget.Toast;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
 
-import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 
-import java.util.concurrent.Executor;
-
-import androidx.annotation.NonNull;
-import androidx.test.rule.ActivityTestRule;
-import app.paseico.data.Router;
-import app.paseico.mainMenu.marketplace.marketplaceFragment;
-
-import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
-import static org.junit.Assert.*;
-
-import java.util.concurrent.Executor;
-
-import app.paseico.data.Router;
-import app.paseico.data.User;
-import app.paseico.login.LogInActivity;
-import app.paseico.mainMenu.marketplace.DiscountsFragment;
 import app.paseico.mainMenu.marketplace.marketplaceFragment;
 
 import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
-public class MarketPlaceTest {
+public class MarketplaceFragmentTest {
 
     @Rule
     public FragmentTestRule<marketplaceFragment> mFragmentTestRule =
@@ -67,7 +39,7 @@ public class MarketPlaceTest {
     }
 
     @AfterClass
-    public void resetPoints(){
+    public static void resetPoints(){
         // Reset the rewardPoints of the user used to test this Fragment.
         DatabaseReference  myActualUserRef = FirebaseDatabase.getInstance().getReference("users").
                 child("7YuiqRHra8OaosZc9vpbfGXdy9C2").child("points");
